refactor: extract extension aggregate queries to core
Move data assembly logic from commands/extension.rs into core
extension module as reusable query functions:

New core APIs:
- extension::list_summaries(project) — pre-computed summaries with
  ready status, compatibility, CLI info, actions, linked status
- extension::update_all(force) — update all non-linked extensions
  with before/after version tracking
- extension::exec_tool(id, component, args) — execute vendor tool
  with enriched PATH and extension env vars

New types in core:
- ExtensionSummary — replaces command-layer ExtensionEntry
- ActionSummary — replaces command-layer duplicate
- UpdateAllResult / UpdateEntry — replaces ExtensionUpdateEntry

commands/extension.rs: 780 → 607 lines (-22%)
- list() is now 6 lines (was 78)
- update_all_extensions() is now 7 lines (was 34)
- exec_extension_tool() is now 9 lines (was 49)
- show_extension() still has manual data assembly (future: core detail())

771 tests pass, cargo fmt clean, zero new warnings.
